requirements.txt
ImageHash==4.3.2
pandas==2.3.3
Pillow==12.1.0
PyExifTool==0.5.6
pytesseract==0.3.13
PyYAML==6.0.3
tqdm==4.67.1
transformers==4.49.0
